move radio to a separate service
