+++ app/src/main/AndroidManifest.xml
@@ -31,5 +31,8 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <service
+            android:name=".RadioPlayerService"
+            android:exported="true" />
     </application>
 </manifest>
